Send TCP RST to client when HTTP/2 error received
As per https://http2.github.io/http2-spec/#CONNECT.

Envoy will send RST_STREAM when MySQL sends a TCP RST, which comes
through on our side as an http2.StreamError.

Currently Envoy is setting the error to NO_ERROR, which is incorrect and
should instead be CONNECT_ERROR. Regardless, the spec states that ANY
error with the stream or connection MUST result in a TCP RST being sent
to the client.